feat: Extracting more functions
